requirements.txt
kafka-python
apache-airflow==2.6.1
cassandra-driver
pyspark
python-dotenv